Improve Python imports
In many cases, modules fields should not be imported directly because
their direct use is ambiguous. For example: `loads` is very hard to
understand whereas `json.loads` is much clearer. Some methods, when used
directly, are also likely to get shadowed accidentally (e.g. `run` from
`subprocess`), potentially causing quite a few headaches.

For some modules and fields, the choice is not obvious. I've tried to
fix the least controversial cases in this commit (e.g. when the cost of
using the module qualifier is low compared to the added clarity in the
code).